+++ app/models/digitization/image.rb
@@ -1,4 +1,6 @@
-class Digitization::Image < ApplicationRecord
+class Digitization::Image < JupiterCore::Depositable
+
+  belongs_to :owner, class_name: 'User'
 
   validates :peel_image_id, uniqueness: true